chore: drop isUsingSyncAPI
- Remove conditional branching based on long dead Gmail rewrite, Inbox
- Consolidate common and gmail page communicators now that the Inbox one has been removed
