+++ src/common/log-error.ts
@@ -13,7 +13,6 @@ export interface LogErrorContext {
   loaderVersion?: string;
   implVersion?: string;
   userEmailHash?: string;
-  isUsingSyncAPI?: boolean;
 }
 
 const sessionId = getSessionId();
@@ -48,7 +47,7 @@ export default function logError(
       err = new Error('Logger.error called with non-error: ' + err);
       markErrorAsSeen(err);
     }
-    const { appIds, implVersion, isUsingSyncAPI, userEmailHash } = context;
+    const { appIds, implVersion, userEmailHash } = context;
     const loaderVersion = context.loaderVersion || BUILD_VERSION;
     const sentByApp = !!context.sentByApp;
 
@@ -86,7 +85,6 @@ export default function logError(
     stuffToLog.push('\nExtension Id:', getExtensionId());
     stuffToLog.push('\nInboxSDK Loader Version:', loaderVersion);
     stuffToLog.push('\nInboxSDK Implementation Version:', implVersion);
-    stuffToLog.push('\nIs Using Sync API:', isUsingSyncAPI);
 
     console.error(...stuffToLog);
 
@@ -102,7 +100,6 @@ export default function logError(
       extensionId: getExtensionId(),
       loaderVersion: loaderVersion,
       implementationVersion: implVersion,
-      isUsingSyncAPI,
       origin: document.location.origin,
       timestamp: Date.now() * 1000,
     };
