+++ src/platform-implementation-js/dom-driver/gmail/gmail-driver/goto-view.ts
@@ -21,21 +21,19 @@ const gotoView = ud.defn(
     }
 
     if (includes(NATIVE_ROUTE_IDS, viewName)) {
-      if (gmailDriver.isUsingSyncAPI()) {
-        // need to go to the new url
-        if (params && typeof params !== 'string') {
-          const threadID = params.threadID;
-          if (threadID && typeof threadID === 'string') {
-            const int = parseInt(threadID, 16);
-            if (!isNaN(int)) {
-              //we got an old id
-              const syncThreadId =
-                await gmailDriver.getSyncThreadIdForOldGmailThreadId(threadID);
-              params = {
-                ...params,
-                threadID: '#' + syncThreadId,
-              };
-            }
+      // need to go to the new url
+      if (params && typeof params !== 'string') {
+        const threadID = params.threadID;
+        if (threadID && typeof threadID === 'string') {
+          const int = parseInt(threadID, 16);
+          if (!isNaN(int)) {
+            //we got an old id
+            const syncThreadId =
+              await gmailDriver.getSyncThreadIdForOldGmailThreadId(threadID);
+            params = {
+              ...params,
+              threadID: '#' + syncThreadId,
+            };
           }
         }
       }
