+++ src/platform-implementation-js/dom-driver/gmail/gmail-page-communicator.ts
@@ -1,15 +1,177 @@
 import { AutocompleteSearchResultWithId } from '../../../injected-js/gmail/modify-suggestions';
-import CommonPageCommunicator from '../../lib/common-page-communicator';
 import { CustomDomEvent } from '../../lib/dom/custom-events';
 import makeMutationObserverChunkedStream from '../../lib/dom/make-mutation-observer-chunked-stream';
 
-import Kefir from 'kefir';
+import Kefir, { type Observable } from 'kefir';
+
+import asap from 'asap';
+import once from 'lodash/once';
+import { CommonAjaxOpts } from '../../../common/ajax';
+import Logger from '../../lib/logger';
 
 // This is intended to be instantiated from makeXhrInterceptor, since it depends
 // on the injected script, and if it's not instantiated elsewhere, you know that
 // if you have an instance of this, then the injected script is present and this
 // will work.
-export default class GmailPageCommunicator extends CommonPageCommunicator {
+export default class GmailPageCommunicator {
+  ajaxInterceptStream: Observable<any, never>;
+
+  constructor() {
+    this.ajaxInterceptStream = Kefir.fromEvents<any, never>(
+      document,
+      'inboxSDKajaxIntercept',
+    ).map((x) => x.detail);
+  }
+
+  getUserEmailAddress(): string {
+    const s = document.head.getAttribute('data-inboxsdk-user-email-address');
+    if (typeof s !== 'string') throw new Error('should not happen');
+    return s;
+  }
+
+  getUserLanguage(): string {
+    const s = document.head.getAttribute('data-inboxsdk-user-language');
+    if (typeof s !== 'string') throw new Error('should not happen');
+    return s;
+  }
+
+  getIkValue(): string {
+    const ownIk = document.head.getAttribute('data-inboxsdk-ik-value');
+    if (ownIk) {
+      return ownIk;
+    }
+    // For standalone windows.
+    if (window.opener) {
+      const openerIk = window.opener.document.head.getAttribute(
+        'data-inboxsdk-ik-value',
+      );
+      if (openerIk) {
+        return openerIk;
+      } else {
+        throw new Error("Failed to look up parent window 'ik' value");
+      }
+    }
+    throw new Error("Failed to look up 'ik' value");
+  }
+
+  async getXsrfToken(): Promise<string> {
+    const existingHeader = document.head.getAttribute(
+      'data-inboxsdk-xsrf-token',
+    );
+    if (existingHeader) {
+      return existingHeader;
+    } else {
+      await this.ajaxInterceptStream
+        .filter(({ type }) => type === 'xsrfTokenHeaderReceived')
+        .take(1)
+        .toPromise();
+
+      const newHeader = document.head.getAttribute('data-inboxsdk-xsrf-token');
+      if (!newHeader) throw new Error('Failed to look up XSRF token');
+      return newHeader;
+    }
+  }
+
+  async getBtaiHeader(): Promise<string> {
+    const existingHeader = document.head.getAttribute(
+      'data-inboxsdk-btai-header',
+    );
+    if (existingHeader) {
+      return existingHeader;
+    } else {
+      await this.ajaxInterceptStream
+        .filter(({ type }) => type === 'btaiHeaderReceived')
+        .take(1)
+        .toPromise();
+
+      const newHeader = document.head.getAttribute('data-inboxsdk-btai-header');
+      if (!newHeader) throw new Error('Failed to look up BTAI header');
+      return newHeader;
+    }
+  }
+
+  resolveUrlRedirects(url: string): Promise<string> {
+    return this.pageAjax({ url, method: 'HEAD' }).then(
+      (result) => result.responseURL,
+    );
+  }
+
+  pageAjax(
+    opts: CommonAjaxOpts,
+  ): Promise<{ text: string; responseURL: string }> {
+    const id = `${Date.now()}-${Math.random()}`;
+    const promise = Kefir.fromEvents(document, 'inboxSDKpageAjaxDone')
+      .filter((event: any) => event.detail && event.detail.id === id)
+      .take(1)
+      .flatMap((event: any) => {
+        if (event.detail.error) {
+          const err = Object.assign(
+            new Error(event.detail.message || 'Connection error') as any,
+            { status: event.detail.status },
+          );
+          if (event.detail.stack) {
+            err.stack = event.detail.stack;
+          }
+          return Kefir.constantError(err);
+        } else {
+          return Kefir.constant({
+            text: event.detail.text,
+            responseURL: event.detail.responseURL,
+          });
+        }
+      })
+      .toPromise();
+
+    document.dispatchEvent(
+      new CustomEvent('inboxSDKpageAjax', {
+        bubbles: false,
+        cancelable: false,
+        detail: Object.assign({}, opts, { id }),
+      }),
+    );
+
+    return promise;
+  }
+
+  silenceGmailErrorsForAMoment(): () => void {
+    document.dispatchEvent(
+      new CustomEvent('inboxSDKsilencePageErrors', {
+        bubbles: false,
+        cancelable: false,
+        detail: null,
+      }),
+    );
+    // create error here for stacktrace
+    const error = new Error('Forgot to unsilence page errors');
+    let unsilenced = false;
+    const unsilence = once(() => {
+      unsilenced = true;
+      document.dispatchEvent(
+        new CustomEvent('inboxSDKunsilencePageErrors', {
+          bubbles: false,
+          cancelable: false,
+          detail: null,
+        }),
+      );
+    });
+    asap(() => {
+      if (!unsilenced) {
+        Logger.error(error);
+        unsilence();
+      }
+    });
+    return unsilence;
+  }
+
+  registerAllowedHashLinkStartTerm(term: string) {
+    document.dispatchEvent(
+      new CustomEvent('inboxSDKregisterAllowedHashLinkStartTerm', {
+        bubbles: false,
+        cancelable: false,
+        detail: { term },
+      }),
+    );
+  }
   async getMessageDate(
     threadId: string,
     message: HTMLElement,
@@ -45,9 +207,10 @@ export default class GmailPageCommunicator extends CommonPageCommunicator {
   ): Promise<T | null> {
     let data = message.getAttribute(attribute);
     if (!data) {
-      const [btaiHeader, xsrfToken] = this.isUsingSyncAPI()
-        ? await Promise.all([this.getBtaiHeader(), this.getXsrfToken()])
-        : [null, null];
+      const [btaiHeader, xsrfToken] = await Promise.all([
+        this.getBtaiHeader(),
+        this.getXsrfToken(),
+      ]);
       message.dispatchEvent(
         new CustomEvent(eventName, {
           bubbles: true,
@@ -134,12 +297,6 @@ export default class GmailPageCommunicator extends CommonPageCommunicator {
     const s = document.head.getAttribute('data-inboxsdk-action-token-value');
     if (s == null) throw new Error('Failed to read value');
     return s;
-  }
-
-  isUsingSyncAPI(): boolean {
-    const s = document.head.getAttribute('data-inboxsdk-using-sync-api');
-    if (s == null) throw new Error('Failed to read value');
-    return s === 'true';
   }
 
   isConversationViewDisabled(): Promise<boolean> {
@@ -264,11 +421,7 @@ export default class GmailPageCommunicator extends CommonPageCommunicator {
     const modifierId = new Date().getTime() + '_' + appId + '_' + Math.random();
 
     const detail: any = { modifierId };
-    if (this.isUsingSyncAPI()) {
-      detail.draftID = keyId;
-    } else {
-      detail.composeid = keyId;
-    }
+    detail.draftID = keyId;
 
     document.dispatchEvent(
       new CustomEvent('inboxSDKregisterComposeRequestModifier', {
@@ -292,9 +445,7 @@ export default class GmailPageCommunicator extends CommonPageCommunicator {
   }
 
   modifyComposeRequest(keyId: string, modifierId: string, composeParams: any) {
-    const detail: any = { modifierId, composeParams };
-    if (this.isUsingSyncAPI()) detail.draftID = keyId;
-    else detail.composeid = keyId;
+    const detail = { modifierId, composeParams, draftID: keyId };
 
     document.dispatchEvent(
       new CustomEvent('inboxSDKcomposeRequestModified', {
